requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
beanie==1.26.0
motor==3.4.0
redis[asyncio]==5.0.4
python-jose[cryptography]==3.3.0
authlib==1.3.0
pydantic-settings==2.2.1
httpx==0.27.0
python-multipart==0.0.9

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
